separate logic, add fun description options
