+++ TODO.md
@@ -3,10 +3,5 @@
 
 -- 
 
-# for each description, create an array of short descriptions 
-# - that get chosen randomly each time the program runs.
-# - - they should be short enough to fit on the screen and range from funny to passive aggressive to helpful
-example: 
- - vim: if youre a super chad -- or something to that effect
- - neovim: does your linked in say "open to work" and you spend all day watching the primeagen?
- - vscode: you're a web developer, aren't you?
+
+
