+++ src/helpers/rando.ts
@@ -1,0 +1,5 @@
+
+
+export function rando(max = 3) { // default is 3, just in case i add more options randomly :D
+  return Math.floor(Math.random() * max);
+}
